requirements.txt
prefect-dbt==0.3.1
anyio==3.7.1
gcsfs==2023.6.0
google-cloud-bigquery==3.11.4
prefect-airbyte==0.2.0
prefect-gcp==0.4.5
prefect-shell==0.1.5
prefect==2.11.5
